fix: standardizes layout of document fields
