+++ packages/payload/src/admin/components/views/Account/Default.tsx
@@ -1,24 +1,19 @@
 import React, { useCallback } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import type { Translation } from '../../../../translations/type'
 import type { CollectionEditViewProps } from '../types'
 
 import { DocumentControls } from '../../elements/DocumentControls'
+import { DocumentFields } from '../../elements/DocumentFields'
 import { DocumentHeader } from '../../elements/DocumentHeader'
-import { Gutter } from '../../elements/Gutter'
 import { LoadingOverlayToggle } from '../../elements/Loading'
-import ReactSelect from '../../elements/ReactSelect'
 import Form from '../../forms/Form'
-import Label from '../../forms/Label'
-import RenderFields from '../../forms/RenderFields'
-import { fieldTypes } from '../../forms/field-types'
 import { LeaveWithoutSaving } from '../../modals/LeaveWithoutSaving'
 import { useAuth } from '../../utilities/Auth'
 import Meta from '../../utilities/Meta'
 import { OperationContext } from '../../utilities/OperationProvider'
 import Auth from '../collections/Edit/Auth'
-import { ToggleTheme } from './ToggleTheme'
+import { Settings } from './Settings'
 import './index.scss'
 
 const baseClass = 'account'
@@ -39,14 +34,7 @@ const DefaultAccount: React.FC<CollectionEditViewProps> = (props) => {
   const { auth, fields } = collection
 
   const { refreshCookieAsync } = useAuth()
-  const { i18n, t } = useTranslation('authentication')
-
-  const languageOptions = Object.entries(i18n.options.resources || {}).map(
-    ([language, resource]) => ({
-      label: (resource as Translation).general.thisLanguage,
-      value: language,
-    }),
-  )
+  const { t } = useTranslation('authentication')
 
   const onSave = useCallback(async () => {
     await refreshCookieAsync()
@@ -55,91 +43,49 @@ const DefaultAccount: React.FC<CollectionEditViewProps> = (props) => {
     }
   }, [onSaveFromProps, refreshCookieAsync])
 
-  const classes = [baseClass].filter(Boolean).join(' ')
-
   return (
     <React.Fragment>
+      <Meta description={t('accountOfCurrentUser')} keywords={t('account')} title={t('account')} />
       <LoadingOverlayToggle name="account" show={isLoading} type="withoutNav" />
       {!isLoading && (
-        <div className={classes}>
-          <OperationContext.Provider value="update">
-            <Form
-              action={action}
-              className={`${baseClass}__form`}
-              disabled={!hasSavePermission}
-              initialState={initialState}
-              method="patch"
-              onSuccess={onSave}
-            >
-              <DocumentHeader apiURL={apiURL} collection={collection} data={data} />
-              <DocumentControls
-                apiURL={apiURL}
-                collection={collection}
-                data={data}
-                hasSavePermission={hasSavePermission}
-                isAccountView
-                permissions={permissions}
-              />
-              <div className={`${baseClass}__main`}>
-                <Meta
-                  description={t('accountOfCurrentUser')}
-                  keywords={t('account')}
-                  title={t('account')}
+        <OperationContext.Provider value="update">
+          <Form
+            action={action}
+            disabled={!hasSavePermission}
+            initialState={initialState}
+            method="patch"
+            onSuccess={onSave}
+          >
+            {!(collection.versions?.drafts && collection.versions?.drafts?.autosave) && (
+              <LeaveWithoutSaving />
+            )}
+            <DocumentHeader apiURL={apiURL} collection={collection} data={data} />
+            <DocumentControls
+              apiURL={apiURL}
+              collection={collection}
+              data={data}
+              hasSavePermission={hasSavePermission}
+              isAccountView
+              permissions={permissions}
+            />
+            <DocumentFields
+              AfterFields={() => <Settings className={`${baseClass}__settings`} />}
+              BeforeFields={() => (
+                <Auth
+                  className={`${baseClass}__auth`}
+                  collection={collection}
+                  email={data?.email}
+                  operation="update"
+                  readOnly={!hasSavePermission}
+                  useAPIKey={auth.useAPIKey}
                 />
-                {!(collection.versions?.drafts && collection.versions?.drafts?.autosave) && (
-                  <LeaveWithoutSaving />
-                )}
-                <div className={`${baseClass}__edit`}>
-                  <Gutter className={`${baseClass}__header`}>
-                    <Auth
-                      className={`${baseClass}__auth`}
-                      collection={collection}
-                      email={data?.email}
-                      operation="update"
-                      readOnly={!hasSavePermission}
-                      useAPIKey={auth.useAPIKey}
-                    />
-                    <RenderFields
-                      fieldSchema={fields}
-                      fieldTypes={fieldTypes}
-                      filter={(field) => field?.admin?.position !== 'sidebar'}
-                      permissions={permissions?.fields}
-                      readOnly={!hasSavePermission}
-                    />
-                  </Gutter>
-                  <Gutter className={`${baseClass}__payload-settings`}>
-                    <h3>{t('general:payloadSettings')}</h3>
-                    <div className={`${baseClass}__language`}>
-                      <Label htmlFor="language-select" label={t('general:language')} />
-                      <ReactSelect
-                        inputId="language-select"
-                        onChange={({ value }) => i18n.changeLanguage(value)}
-                        options={languageOptions}
-                        value={languageOptions.find((language) => language.value === i18n.language)}
-                      />
-                    </div>
-                    <ToggleTheme />
-                  </Gutter>
-                </div>
-              </div>
-              <div className={`${baseClass}__sidebar-wrap`}>
-                <div className={`${baseClass}__sidebar`}>
-                  <div className={`${baseClass}__sidebar-sticky-wrap`}>
-                    <div className={`${baseClass}__sidebar-fields`}>
-                      <RenderFields
-                        fieldSchema={fields}
-                        fieldTypes={fieldTypes}
-                        filter={(field) => field?.admin?.position === 'sidebar'}
-                        permissions={permissions?.fields}
-                        readOnly={!hasSavePermission}
-                      />
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </Form>
-          </OperationContext.Provider>
-        </div>
+              )}
+              fields={fields}
+              hasSavePermission={hasSavePermission}
+              permissions={permissions}
+            />
+          </Form>
+        </OperationContext.Provider>
       )}
     </React.Fragment>
   )
