+++ packages/payload/src/admin/components/views/Account/Settings/index.scss
@@ -1,0 +1,45 @@
+@import '../../../../scss/styles.scss';
+
+.payload-settings {
+  position: relative;
+
+  h3 {
+    margin: 0;
+  }
+
+  &::before,
+  &::after {
+    content: '';
+    display: block;
+    height: 1px;
+    background: var(--theme-elevation-100);
+    width: calc(100% + calc(var(--base) * 5));
+    left: calc(var(--gutter-h) * -1);
+    top: 0;
+    position: absolute;
+  }
+
+  &::after {
+    display: none;
+    bottom: 0;
+    top: unset;
+  }
+
+  margin-top: base(3);
+  padding-top: base(3);
+  padding-bottom: base(3);
+  display: flex;
+  flex-direction: column;
+  gap: var(--base);
+
+  @include mid-break {
+    margin-bottom: var(--base);
+    padding-top: base(2);
+    margin-top: base(2);
+    padding-bottom: base(2);
+
+    &::after {
+      display: block;
+    }
+  }
+}
