+++ packages/payload/src/admin/components/views/collections/Edit/Default/index.scss
@@ -1,168 +1,15 @@
 @import '../../../../../scss/styles.scss';
 
 .collection-default-edit {
-  width: 100%;
-  display: flex;
-  --doc-sidebar-width: 500px;
-
-  &--has-sidebar {
-    .collection-default-edit {
-      &__edit {
-        [dir='ltr'] & {
-          top: 0;
-          right: 0;
-          border-right: 1px solid var(--theme-elevation-100);
-          padding-right: calc(var(--base) * 2);
-        }
-
-        [dir='rtl'] & {
-          top: 0;
-          left: 0;
-          border-left: 1px solid var(--theme-elevation-100);
-          padding-left: calc(var(--base) * 2);
-        }
-      }
-
-      &__fields {
-        & > .tabs-field,
-        & > .group-field {
-          margin-right: calc(var(--base) * -2);
-        }
-      }
-    }
-  }
-
-  &__main {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    min-height: 100%;
-    flex-grow: 1;
-  }
-
-  &__edit {
-    padding-top: calc(var(--base) * 1.5);
-    padding-bottom: var(--spacing-view-bottom);
-    flex-grow: 1;
-  }
-
   &__auth {
-    margin-bottom: var(--base);
-  }
-
-  &__sidebar-wrap {
-    position: sticky;
-    top: var(--doc-controls-height);
-    height: calc(100vh - var(--doc-controls-height));
-    width: var(--doc-sidebar-width);
-    flex-shrink: 0;
-  }
-
-  &__sidebar {
-    width: 100%;
-    height: 100%;
-    overflow-y: auto;
-  }
-
-  &__sidebar-sticky-wrap {
-    display: flex;
-    flex-direction: column;
-    min-height: 100%;
-  }
-
-  &__sidebar-fields {
-    display: flex;
-    flex-direction: column;
-    gap: var(--base);
-    padding-top: calc(var(--base) * 1.5);
-    padding-left: calc(var(--base) * 2);
-    padding-right: var(--gutter-h);
-    padding-bottom: var(--spacing-view-bottom);
-  }
-
-  &__label {
-    color: var(--theme-elevation-400);
-  }
-
-  @include large-break {
-    --doc-sidebar-width: 350px;
-  }
-
-  @include mid-break {
-    display: block;
-
-    &--has-sidebar {
-      .collection-default-edit {
-        &__main {
-          width: 100%;
-        }
-
-        &__edit {
-          [dir='ltr'] & {
-            border-right: 0;
-            padding-right: var(--gutter-h);
-          }
-
-          [dir='rtl'] & {
-            border-left: 0;
-            padding-left: var(--gutter-h);
-          }
-        }
-
-        &__fields {
-          & > .tabs-field,
-          & > .group-field {
-            margin-right: calc(var(--gutter-h) * -1);
-          }
-        }
-      }
-    }
-
-    &__main {
-      width: 100%;
-      min-height: initial;
-    }
-
-    &__sidebar-wrap {
-      position: static;
-      width: 100%;
-      height: initial;
-      border-left: 0;
-      margin-top: calc(var(--base) / 2);
-    }
-
-    &__form {
-      display: block;
-    }
-
-    &__sidebar-fields {
-      padding-top: 0;
-      padding-left: var(--gutter-h);
-      padding-right: var(--gutter-h);
-      gap: base(0.5);
-
-      [dir='ltr'] & {
-        padding-right: var(--gutter-h);
-      }
-
-      [dir='rtl'] & {
-        padding-left: var(--gutter-h);
-      }
-    }
-
-    &__sidebar {
-      padding-bottom: base(3.5);
-      overflow: visible;
-    }
+    margin-bottom: calc(var(--base) * 2);
+    margin-top: calc(var(--base) * 0.5);
   }
 
   @include small-break {
-    &__sidebar-wrap {
-      min-width: initial;
-    }
-
-    &__edit {
-      padding-top: calc(var(--base) / 2);
+    &__auth {
+      margin-top: 0;
+      margin-bottom: var(--base);
     }
   }
 }
