+++ packages/payload/src/admin/components/views/collections/Edit/Default/index.tsx
@@ -5,10 +5,7 @@ import type { CollectionEditViewProps } from '../../../types'
 
 import { getTranslation } from '../../../../../../utilities/getTranslation'
 import { DocumentControls } from '../../../../elements/DocumentControls'
-import { Gutter } from '../../../../elements/Gutter'
-import RenderFields from '../../../../forms/RenderFields'
-import { filterFields } from '../../../../forms/RenderFields/filterFields'
-import { fieldTypes } from '../../../../forms/field-types'
+import { DocumentFields } from '../../../../elements/DocumentFields'
 import { LeaveWithoutSaving } from '../../../../modals/LeaveWithoutSaving'
 import Meta from '../../../../utilities/Meta'
 import Auth from '../Auth'
@@ -38,18 +35,21 @@ export const DefaultCollectionEdit: React.FC<CollectionEditViewProps> = (props) 
 
   const operation = isEditing ? 'update' : 'create'
 
-  const sidebarFields = filterFields({
-    fieldSchema: fields,
-    fieldTypes,
-    filter: (field) => field?.admin?.position === 'sidebar',
-    permissions: permissions.fields,
-    readOnly: !hasSavePermission,
-  })
-
-  const hasSidebar = sidebarFields && sidebarFields.length > 0
-
   return (
     <Fragment>
+      <Meta
+        description={`${isEditing ? t('editing') : t('creating')} - ${getTranslation(
+          collection.labels.singular,
+          i18n,
+        )}`}
+        keywords={`${getTranslation(collection.labels.singular, i18n)}, Payload, CMS`}
+        title={`${isEditing ? t('editing') : t('creating')} - ${getTranslation(
+          collection.labels.singular,
+          i18n,
+        )}`}
+      />
+      {!(collection.versions?.drafts && collection.versions?.drafts?.autosave) &&
+        !disableLeaveWithoutSaving && <LeaveWithoutSaving />}
       <SetStepNav collection={collection} id={id} isEditing={isEditing} />
       <DocumentControls
         apiURL={apiURL}
@@ -61,29 +61,9 @@ export const DefaultCollectionEdit: React.FC<CollectionEditViewProps> = (props) 
         isEditing={isEditing}
         permissions={permissions}
       />
-      <div
-        className={[
-          baseClass,
-          hasSidebar ? `${baseClass}--has-sidebar` : `${baseClass}--no-sidebar`,
-        ]
-          .filter(Boolean)
-          .join(' ')}
-      >
-        <div className={`${baseClass}__main`}>
-          <Meta
-            description={`${isEditing ? t('editing') : t('creating')} - ${getTranslation(
-              collection.labels.singular,
-              i18n,
-            )}`}
-            keywords={`${getTranslation(collection.labels.singular, i18n)}, Payload, CMS`}
-            title={`${isEditing ? t('editing') : t('creating')} - ${getTranslation(
-              collection.labels.singular,
-              i18n,
-            )}`}
-          />
-          {!(collection.versions?.drafts && collection.versions?.drafts?.autosave) &&
-            !disableLeaveWithoutSaving && <LeaveWithoutSaving />}
-          <Gutter className={`${baseClass}__edit`}>
+      <DocumentFields
+        BeforeFields={() => (
+          <Fragment>
             {auth && (
               <Auth
                 className={`${baseClass}__auth`}
@@ -97,33 +77,12 @@ export const DefaultCollectionEdit: React.FC<CollectionEditViewProps> = (props) 
               />
             )}
             {upload && <Upload collection={collection} internalState={internalState} />}
-            <RenderFields
-              className={`${baseClass}__fields`}
-              fieldSchema={fields}
-              fieldTypes={fieldTypes}
-              filter={(field) => !field?.admin?.position || field?.admin?.position !== 'sidebar'}
-              permissions={permissions.fields}
-              readOnly={!hasSavePermission}
-            />
-          </Gutter>
-        </div>
-        {hasSidebar && (
-          <div className={`${baseClass}__sidebar-wrap`}>
-            <div className={`${baseClass}__sidebar`}>
-              <div className={`${baseClass}__sidebar-sticky-wrap`}>
-                <div className={`${baseClass}__sidebar-fields`}>
-                  <RenderFields
-                    fieldTypes={fieldTypes}
-                    fields={sidebarFields}
-                    permissions={permissions.fields}
-                    readOnly={!hasSavePermission}
-                  />
-                </div>
-              </div>
-            </div>
-          </div>
+          </Fragment>
         )}
-      </div>
+        fields={fields}
+        hasSavePermission={hasSavePermission}
+        permissions={permissions}
+      />
     </Fragment>
   )
 }
